chore: refactor route files into individual folders for modularity and separation of concerns
