+++ src/App.js
@@ -6,18 +6,18 @@ import '@fontsource/roboto/400.css';
 import '@fontsource/roboto/500.css';
 import '@fontsource/roboto/700.css';
 
-import Home from './Routes/Home';
-import FetchAllBookings from './Routes/FetchAllBookings';
-import FetchAllRooms from './Routes/FetchAllRooms';
-import GetRoomDetails from './Routes/GetRoomDetails';
-import BookHotel from './Routes/BookHotel';
-import Help from './Routes/Help';
-import ContactUsHelp from './Routes/ContactUsHelp';
-import AboutHelp from './Routes/AboutHelp';
-import FrequentlyAskedQuestionsHelp from './Routes/FrequentlyAskedQuestionsHelp';
-import Login from './Routes/Login';
-import SignUp from './Routes/SignUp';
-import NavbarComponent from './Routes/NavBar';
+import Home from './Routes/Home/Home';
+import FetchAllBookings from './Routes/AllBookings/FetchAllBookings';
+import FetchAllRooms from './Routes/AllHotels/FetchAllRooms';
+import GetRoomDetails from './Routes/HotelDetails/GetRoomDetails';
+import BookHotel from './Routes/BookHotel/BookHotel';
+import Help from './Routes/Help/Help';
+import ContactUsHelp from './Routes/ContactUsHelp/ContactUsHelp';
+import AboutHelp from './Routes/AboutUsHelp/AboutHelp';
+import FrequentlyAskedQuestionsHelp from './Routes/FAQs/FrequentlyAskedQuestionsHelp';
+import Login from './Routes/Login/Login';
+import SignUp from './Routes/Register/SignUp';
+import NavbarComponent from './Routes/Navbar/NavBar';
 import PaymentSuccess from './Routes/PaymentSuccess';
 import PaymentCanceled from './Routes/PaymentCanceled';
