+++ src/Routes/ContactUsHelp/ContactUsHelp.js
@@ -5,8 +5,8 @@ import Form from 'react-bootstrap/Form';
 import FloatingLabel from 'react-bootstrap/FloatingLabel';
 import Button from 'react-bootstrap/Button';
 
-import Footer from './Footer';
-import NavbarComponent from './NavBar';
+import Footer from '../Footer/Footer';
+import NavbarComponent from '../Navbar/NavBar';
 
 const ContactUsHelp = () => {
   // component always displays from top on initial render.
